requirements.txt
mcp==1.25.0
fastmcp==2.14.1
pydantic-ai==1.37.0
requests==2.31.0
python-dotenv==1.0.0